Refactor: rename record_util.py
