+++ src/caret_analyze/record/__init__.py
@@ -16,9 +16,9 @@ from .column import Column, Columns, ColumnValue
 from .data_frame_shaper import Clip, DataFrameShaper, Strip
 from .interface import RecordInterface, RecordsInterface
 from .record_factory import RecordFactory, RecordsFactory
-from .record_utils import (merge,
-                           merge_sequential,
-                           merge_sequential_for_addr_track)
+from .record_operations import (merge,
+                                merge_sequential,
+                                merge_sequential_for_addr_track)
 from .records_service import Frequency, Latency, Period, Range, ResponseTime, StackedBar
 
 __all__ = [
