+++ src/caret_analyze/runtime/path.py
@@ -24,8 +24,8 @@ from .path_base import PathBase
 from ..common import Summarizable, Summary, Util
 from ..exceptions import Error, InvalidArgumentError, InvalidRecordsError
 from ..record import Columns, RecordsFactory
-from ..record.record_utils import merge, merge_sequential
 from ..record.interface import RecordsInterface
+from ..record.record_operations import merge, merge_sequential
 from ..value_objects import CallbackChain, PathStructValue
 
 logger = getLogger(__name__)
